% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_util.R
\name{append_big.dat_parquet}
\alias{append_big.dat_parquet}
\title{append_big.dat_parquet}
\usage{
append_big.dat_parquet(
  mat,
  big.dat,
  col.bin.size = 50000,
  row.bin.size = 500,
  mc.cores = 10
)
}
\arguments{
\item{mat}{}

\item{big.dat}{}

\item{col.bin.size}{}

\item{row.bin.size}{}
}
\value{
value
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
