% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_merge.R
\name{combine_cl_sim}
\alias{combine_cl_sim}
\title{combine_cl_sim}
\usage{
combine_cl_sim(cl.rd.list, cl, comb.dat)
}
\arguments{
\item{cl.rd.list}{}

\item{cl}{}

\item{comb.dat}{}
}
\value{
value.
}
\description{
##' .. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
