% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_merge.R
\name{de_genes_pairs_multiple}
\alias{de_genes_pairs_multiple}
\title{Title: de_genes_pairs_multiple}
\usage{
de_genes_pairs_multiple(
  dat.list,
  de.param.list,
  cl,
  pairs,
  cl.means.list,
  cl.present.list,
  cl.sqr.means.list,
  lfc.conservation.th = 0.6,
  de.genes.list = NULL,
  max.cl.size = 200,
  mc.cores = 1,
  method = "fast_limma"
)
}
\arguments{
\item{dat.list}{dat.list}

\item{de.param.list}{de.param.list}

\item{cl}{cl}

\item{pairs}{pairs}

\item{cl.means.list}{cl.means.list}

\item{cl.present.list}{cl.present.list}

\item{lfc.conservation.th}{lfc.conservation.th}

\item{de.genes.list}{de.genes.list}

\item{max.cl.size}{max.cl.size}

\item{common.genes}{common.genes}
}
\value{
value.
}
\description{
Title: de_genes_pairs_multiple
}
