% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_pair_t.test}
\alias{de_pair_t.test}
\title{Perform pairwise differential detection tests using t.test for a single pair of clusters}
\usage{
de_pair_t.test(pair, cl.means, cl.present, cl.vars, cl.size)
}
\arguments{
\item{pair}{a numeric vector of length 2 specifying which clusters to compare}

\item{cl.means}{a data.frame of normalized mean gene expression values (genes x clusters)}

\item{cl.vars}{a data.frame of normalized variance gene expression values (genes x clusters)}

\item{cl.size}{a named numeric vector of cluster sizes}

\item{genes}{the genes to use for pairwise comparisons}
}
\value{
value.  a data.frame with DE statistics:
\itemize{
\item{padj} P-values adjusted using the Holm (1979) method (\code{p.adjust()} default).
\item{pval} P-values reported by the \code{vec_chisq_test()} function.
\item{lfc} Log fold change of mean expression values between the pair of clusters.
\item{meanA} Mean expression value for the first cluster in the pair.
\item{meanB} Mean expression value for the second cluster in the pair.
\item{q1} Proportion of cells expressing each gene for the first cluster in the pair.
\item{q2} Proportion of cells expressing each gene for the second cluster in the pair.
}
}
\description{
Perform pairwise differential detection tests using t.test for a single pair of clusters
}
