% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cl_stats}
\alias{get_cl_stats}
\title{Perform matrix operation using RCPP}
\usage{
get_cl_stats(
  mat,
  cl,
  stats = c("sums", "means", "medians", "present", "sqr_sums", "sqr_means"),
  low.th = 1,
  parallel = FALSE,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{mat}{a numeric matrix}

\item{cl}{a cluster factor object}

\item{stats}{operation string. Currently support: sums, means, median, present, sqr_means}

\item{parallel}{boolean to indicate if running the rcpp function in parallel or not}

\item{sparse}{boolean] to indicate if the matrix is sparse or not}

\item{transpose}{boolean to indicate if the matrix is transposed or not}

\item{numThreads}{integer to indicate the number of threads}
}
\value{
the resulting numeric matrix after applying the operation
}
\description{
Currently supported operations are: sums, means, median, present, sqr_means
}
