% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_gene_score}
\alias{get_gene_score}
\title{Title}
\usage{
get_gene_score(
  de.genes,
  cl.means = NULL,
  all.genes = NULL,
  top.n = 50,
  max.num = 1000,
  bin.th = 4,
  mc.cores = 1
)
}
\arguments{
\item{de.genes}{}

\item{top.n}{}

\item{max.num}{}

\item{bin.th}{}
}
\value{
value.
}
\description{
Title
}
