% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{get_knn}
\alias{get_knn}
\title{Get KNN}
\usage{
get_knn(
  dat,
  ref.dat,
  k,
  method = "cor",
  dim = NULL,
  index = NULL,
  build.index = FALSE,
  transposed = TRUE,
  return.distance = FALSE,
  ntrees = 100
)
}
\arguments{
\item{dat}{}

\item{ref.dat}{}

\item{k}{}

\item{method}{}

\item{dim}{}
}
\value{
value.
}
\description{
Get KNN
}
