% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.graph.R
\name{get_knn_graph}
\alias{get_knn_graph}
\title{get knn graph}
\usage{
get_knn_graph(
  rd.dat,
  cl,
  ref.cells = row.names(rd.dat),
  method = "Annoy.Cosine",
  k = 15,
  knn.outlier.th = 2,
  outlier.frac.th = 0.5,
  clean.cells = row.names(rd.dat),
  knn.result = NULL,
  mc.cores = 10
)
}
\arguments{
\item{rd.dat}{}

\item{cl}{}

\item{k}{}

\item{knn.outlier.th}{}

\item{outlier.frac.th}{}
}
\value{
value.
}
\description{
get knn graph
}
