% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{get_RD_cl_center}
\alias{get_RD_cl_center}
\title{get_RD_cl_center}
\usage{
get_RD_cl_center(rd.dat, cl)
}
\arguments{
\item{rd.dat}{}

\item{cl}{}
}
\value{
value.
}
\description{
get_RD_cl_center
}
