% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{i_harmonize}
\alias{i_harmonize}
\title{i_harmonize}
\usage{
i_harmonize(
  comb.dat,
  select.cells = comb.dat$all.cells,
  ref.sets = names(comb.dat$dat.list),
  select.sets = names(comb.dat$dat.list),
  prefix = "",
  joint.rd.dat = NULL,
  result = NULL,
  overwrite = TRUE,
  sample.size = 50000,
  dir = "./",
  rm.genes = NULL,
  split.size = 100,
  ...
)
}
\arguments{
\item{comb.dat}{}

\item{select.cells}{}

\item{prefix}{}

\item{result}{}

\item{...}{}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
