% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_big.R
\name{iter_clust_big}
\alias{iter_clust_big}
\title{Iterative clustering algorithm for single cell RNAseq dataset}
\usage{
iter_clust_big(
  big.dat = NULL,
  select.cells = big.dat$col_id,
  prefix = NULL,
  split.size = 10,
  result = NULL,
  method = "auto",
  counts = NULL,
  sampleSize = 50000,
  mc.cores = 10,
  overwrite = TRUE,
  verbose = FALSE,
  jaccard.sampleSize = 3e+05,
  ...
)
}
\arguments{
\item{select.cells}{The cells to be clustered}

\item{prefix}{The character string to indicate current iteration.}

\item{split.size}{The minimal cluster size for further splitting}

\item{result}{The current clustering result as basis for further splitting.}

\item{method}{Clustering method. It can be "auto", "louvain", "hclust"}

\item{...}{Other parameters passed to method `onestep_clust()`}

\item{norm.dat}{normalized expression data matrix in log transform, using genes as rows, and cells and columns. Users can use log2(FPKM+1) or log2(CPM+1)}
}
\value{
Clustering result is returned as a list with two elements:
        cl: cluster membership for each cell
        markers: top markers that seperate clusters
}
\description{
Iterative clustering algorithm for single cell RNAseq dataset
}
\examples{
clust.result <- iter_clust(norm.dat)
          clust.result <- iter_clust(norm.dat, de.param = de_param(q1.th = 0.5, de.score.th = 100))
}
