% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{jaccard_big}
\alias{jaccard_big}
\title{jaccard_big}
\usage{
jaccard_big(knn.index, bin_size = 50000, prune = 0)
}
\arguments{
\item{knn.index}{}

\item{bin_size}{}

\item{prune}{}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
