% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{knn_jaccard_clust}
\alias{knn_jaccard_clust}
\title{knn_jaccard_clust}
\usage{
knn_jaccard_clust(
  knn.index,
  method = c("leiden", "louvain"),
  prune = 0.05,
  return.graph = FALSE,
  ...
)
}
\arguments{
\item{knn.index}{}

\item{method}{}

\item{prune}{}
}
\value{
value.
}
\description{
KNN Jaccard Louvain
}
\author{
Zizhen Yao
}
