% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{knn_joint}
\alias{knn_joint}
\title{knn_joint}
\usage{
knn_joint(
  comb.dat,
  joint.rd.dat = NULL,
  ref.sets = names(comb.dat$dat.list),
  select.sets = names(comb.dat$dat.list),
  merge.sets = ref.sets,
  select.cells = comb.dat$all.cells,
  select.genes = NULL,
  genes.allowed = NULL,
  cross.knn.method = "Annoy.Cosine",
  self.knn.method = "Annoy.Euclidean",
  method = "leiden",
  k = 15,
  sample.size = 50000,
  jaccard.sampleSize = 2e+05,
  cl.sample.size = 100,
  block.size = 10000,
  verbose = TRUE,
  mc.cores = 1,
  rm.eigen = NULL,
  rm.th = 0.7,
  max.dim = 20,
  rm.genes = NULL,
  ...
)
}
\arguments{
\item{comb.dat}{}

\item{ref.sets}{}

\item{select.sets}{}

\item{merge.sets}{}

\item{select.cells}{}

\item{select.genes}{}

\item{cross.knn.method}{}

\item{self.knn.method}{}

\item{k}{}

\item{sample.size}{}

\item{cl.sample.size}{}

\item{block.size}{}

\item{verbose}{}

\item{mc.cores}{}

\item{rm.eigen}{}

\item{rm.th}{}

\item{...}{}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
