% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_merge.R
\name{merge_cl_multiple}
\alias{merge_cl_multiple}
\title{merge_cl_multiple}
\usage{
merge_cl_multiple(
  comb.dat,
  merge.sets,
  cl,
  anchor.genes = NULL,
  genes.allowed = NULL,
  joint.rd.dat = NULL,
  verbose = TRUE,
  lfc.conservation.th = 0.7,
  merge.type = "undirectional",
  de.method = "fast_limma",
  max.cl.size = 300,
  compare.k = 4,
  mc.cores = 5,
  pairBatch = 100,
  cl.stats.list = NULL
)
}
\arguments{
\item{comb.dat}{comb.dat}

\item{merge.sets}{merge.sets}

\item{cl}{cl}

\item{anchor.genes}{anchor.genes}

\item{verbose}{verbose}

\item{lfc.conservation.th}{lfc.conservation.th}

\item{merge.type}{merge.type}

\item{de.method}{de.method}

\item{max.cl.size}{max.cl.size}

\item{compare.k}{compare.k}

\item{mc.cores}{mc.cores}

\item{pairBatch}{pairBatch}

\item{cl.stats.list}{cl.stats.list}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
