% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{multiplot}
\alias{multiplot}
\title{Title}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL, byrow = FALSE)
}
\arguments{
\item{...}{}

\item{plotlist}{}

\item{cols}{}

\item{layout}{}

\item{file}{}
}
\value{
value.
}
\description{
Title
}
