% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_2d_umap_anno}
\alias{plot_2d_umap_anno}
\title{plot_2d_umap_anno}
\usage{
plot_2d_umap_anno(
  umap.fn,
  anno.df,
  dest.d = "./",
  umap.2d = NULL,
  meta.fields = NULL,
  show.label = FALSE,
  alpha = 0.65,
  cex = 0.25,
  save.format = c("png", "pdf", "both"),
  plot.height = 7,
  plot.width = 7,
  show.legend = FALSE
)
}
\arguments{
\item{umap.fn}{path to umap coordinates. CSV file containing sample_id and umap x/y coordinates}

\item{anno.df}{Sample annotations. The first column should be sample_id, and each annotation should have \_id, \_label, and \_color columns. Requires cluster_id which needs to be sequential in order of the dendrogram.}

\item{dest.d}{path to save plots.}

\item{meta.fields}{base name of variables to be represented as bargraphs below dendrogram. Annotation variables need to be represented as \_id, \_label, \_color in anno.}

\item{show.label}{TRUE or FALSE. To show cluster label on top of plot.}

\item{alpha}{level of transparency of dots. Value between 0 (fully transparent) and 1 (fully opaque)}

\item{cex}{size of plotted points. Default = 0.25}

\item{save.format}{figures can be saved as "png", "pdf" or "both"}

\item{plot.height}{plot.height}

\item{plot.width}{plot.width}

\item{show.legend}{TRUE or FALSE. Whether to show legend with plot.}
}
\description{
plot_2d_umap_anno
}
\examples{
# First, load the example data
load("data/rd_plot_example/example_umap.csv")
load("data/rd_plot_example/anno.df.rda")

# Then, run the function with the loaded data
plots <- plot_2d_umap_anno(umap.fn = "data/rd_plot_example/example_umap.csv",
                           anno.df = anno.df, dest.d = "./",
                           meta.fields = c("platform", "joint_region"),
                           show.label = FALSE, alpha = 0.5, cex = 0.15,
                           save.format = "both", plot.height = 7, plot.width = 10,
                           show.legend = TRUE)h=10, show.legend=TRUE)

}
