% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_RD_cl}
\alias{plot_RD_cl}
\title{plot_RD_cl}
\usage{
plot_RD_cl(
  rd.dat,
  cl,
  cl.color = NULL,
  cl.label = NULL,
  cex = 0.15,
  fn.size = 2,
  alpha.val = NULL,
  show.legend = FALSE,
  legend.size = 2,
  label.center = TRUE,
  bg = "blank",
  fn.color = "black",
  na.val = "grey60",
  no.shape = TRUE,
  ncol = 4,
  shift.x = 0,
  shift.y = 0,
  raster = F,
  dpi = 300
)
}
\arguments{
\item{rd.dat}{}

\item{cl}{}

\item{cl.color}{}

\item{cl.label}{}

\item{cex}{}

\item{fn.size}{}

\item{alpha.val}{}

\item{show.legend}{}

\item{legend.size}{}

\item{label.center}{}

\item{bg}{}

\item{fn.color}{}

\item{raster}{whether to rasterize the scatterplot. Default=F}
}
\value{
value.
}
\description{
plot_RD_cl
}
