% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_RD_cl_subset}
\alias{plot_RD_cl_subset}
\title{plot_RD_cl_subset}
\usage{
plot_RD_cl_subset(
  rd.dat,
  cl,
  cl.color,
  cl.label,
  select.samples,
  missing.color = "gray85",
  min.size = 10,
  fg.alpha = 1,
  bg.alpha = 0.5,
  fg.size = 1,
  bg.size = 0.15,
  ...
)
}
\arguments{
\item{rd.dat}{}

\item{cl}{}

\item{cl.color}{}

\item{cl.label}{}

\item{select.samples}{}

\item{missing.color}{}

\item{min.size}{}

\item{fg.alpha}{}

\item{bg.alpha}{}

\item{fg.size}{}

\item{bg.size}{}

\item{...}{}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
