% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_RD_gene}
\alias{plot_RD_gene}
\title{plot_RD_gene}
\usage{
plot_RD_gene(rd.dat, norm.dat, genes, cex = 0.15, raster = F, dpi = 300)
}
\arguments{
\item{rd.dat}{}

\item{norm.dat}{}

\item{genes}{}

\item{cex}{}
}
\value{
value.
}
\description{
plot_RD_gene
}
