% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_RD_meta}
\alias{plot_RD_meta}
\title{plot_RD_meta}
\usage{
plot_RD_meta(
  rd.dat,
  meta,
  meta.col = NULL,
  show.legend = TRUE,
  cex = 0.15,
  legend.size = 5,
  alpha.val = 1,
  palette = NULL,
  reverse = FALSE,
  low = "blue",
  high = "red",
  raster = F,
  dpi = 300,
  na.val = "grey60"
)
}
\arguments{
\item{rd.dat}{}

\item{meta}{}

\item{meta.col}{}

\item{show.legend}{}

\item{cex}{}

\item{legend.size}{}

\item{alpha.val}{}

\item{palette}{default it 2 color blue-red but can be defined using low and high params. Preset options include c("comet","greymagma","ylgnbu", "spectral", "turbo" )}

\item{raster}{whether to rasterize the scatterplot. Default=F}
}
\value{
value.
}
\description{
plot_RD_meta
}
