% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_impute.R
\name{predict_knn}
\alias{predict_knn}
\title{predict_knn}
\usage{
predict_knn(knn.idx, reference, cl, mc.cores = 1)
}
\arguments{
\item{knn.idx}{}

\item{reference}{}

\item{cl}{}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
