% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_markers_groups}
\alias{select_markers_groups}
\title{Title}
\usage{
select_markers_groups(
  de.dir,
  cl.group,
  genes,
  cl.bin,
  select.groups = unique(cl.group$group),
  lfc.th = 2,
  cl.means = NULL,
  n.markers = 20,
  ds = NULL,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{cl.group}{Assignment of clusters to groups cluster as names, and group id as values.}

\item{de.genes}{}
}
\value{
value.
}
\description{
Title
}
