% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.parquet.R
\name{select_markers_groups_top_ds}
\alias{select_markers_groups_top_ds}
\title{Title}
\usage{
select_markers_groups_top_ds(
  ds,
  cl.group,
  select.groups = names(cl.group),
  n.markers = 3,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{cl.group}{Assignment of clusters to groups cluster as names, and group id as values.}

\item{de.genes}{}
}
\value{
value.
}
\description{
Title
}
