% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.parquet.R
\name{select_markers_pair_direction_ds}
\alias{select_markers_pair_direction_ds}
\title{select_markers_pair_direction_ds}
\usage{
select_markers_pair_direction_ds(
  de.dir,
  add.num,
  genes,
  cl.bin,
  de = NULL,
  mc.cores = 20,
  max.genes = 1000,
  cl.means = NULL,
  lfc.th = 2,
  ds = NULL
)
}
\arguments{
\item{de.dir}{}

\item{add.num}{}

\item{genes}{}

\item{cl.bin}{}

\item{de}{}

\item{mc.cores}{}

\item{max.genes}{}

\item{cl.means}{}

\item{lfc.th}{}
}
\value{
value.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Zizhen Yao
}
