% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_markers_pair_group_top}
\alias{select_markers_pair_group_top}
\title{Title}
\usage{
select_markers_pair_group_top(
  cl,
  g1,
  g2,
  de.genes,
  cl.means,
  top.n = 50,
  max.num = 1000,
  n.markers = 20,
  up.gene.score = NULL,
  down.gene.score = NULL
)
}
\arguments{
\item{cl}{}

\item{g1}{}

\item{g2}{}

\item{de.genes}{}

\item{top.n}{}

\item{max.num}{}

\item{n.markers}{}

\item{up.gene.score}{}

\item{down.gene.score}{}
}
\value{
value.
}
\description{
Title
}
