% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.parquet.R
\name{select_markers_pair_group_top_ds}
\alias{select_markers_pair_group_top_ds}
\title{Title: select_markers_pair_group_top_ds}
\usage{
select_markers_pair_group_top_ds(
  g1,
  g2,
  ds,
  genes,
  cl.bin,
  select.sign = "up",
  n.markers = 20,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{g1}{g1}

\item{g2}{g2}

\item{ds}{ds}

\item{n.markers}{n.markers}

\item{cl}{cl}

\item{top.n}{top.n}

\item{max.num}{max.num}

\item{up.gene.score}{up.gene.score}

\item{down.gene.score}{down.gene.score}
}
\value{
value. Note # Always directory
}
\description{
Title: select_markers_pair_group_top_ds
}
