% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.parquet.R
\name{select_pos_markers_ds}
\alias{select_pos_markers_ds}
\title{Title}
\usage{
select_pos_markers_ds(
  de.dir,
  cl,
  select.cl,
  genes,
  cl.bin,
  ds = NULL,
  n.markers = 1,
  out.dir = "cl.markers",
  mc.cores = 1,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{cl}{}

\item{n.markers}{}

\item{de.genes}{}

\item{default.markers}{}

\item{rm.genes}{}

\item{up.gene.score}{}

\item{down.gene.score}{}
}
\value{
value.
}
\description{
Title
}
